noticed the new instanciation was not the right way to erase a single object and display it. an if statement instead triggered the change to appear

+++ MainApp_WPF_Presentation/ViewModels/ContactDisplayViewModel.cs
@@ -36,7 +36,11 @@ public partial class ContactDisplayViewModel : ObservableObject
     [RelayCommand]
     private void DeleteOne(Contact contact)
     {
-        _contactService.RemoveOne(contact);
-        Contacts = new ObservableCollection<Contact>();
+
+        if (Contacts.Contains(contact))
+        {
+            _contactService.RemoveOne(contact);
+            Contacts.Remove(contact);
+        }
     }
 }
